tests(firebase_storage): use e2e
