--- packages/firebase_storage/example/test_driver/firebase_storage_test.dart
@@ -1,7 +1,0 @@
-import 'package:flutter_driver/flutter_driver.dart';
-
-Future<void> main() async {
-  final FlutterDriver driver = await FlutterDriver.connect();
-  await driver.requestData(null, timeout: const Duration(minutes: 1));
-  driver.close();
-}
